refactor(front): move whitelisted-model selectors to lib/api/assistant/models
