+++ front/lib/api/assistant/global_agents/configurations/dust/dust.ts
@@ -24,11 +24,11 @@ import {
   _getToolsetsToolsConfiguration,
 } from "@app/lib/api/assistant/global_agents/tools";
 import { dummyModelConfiguration } from "@app/lib/api/assistant/global_agents/utils";
-import { isProviderWhitelisted } from "@app/lib/api/assistant/models";
 import {
-  getLargeWhitelistedModel,
   getSmallWhitelistedModel,
-} from "@app/lib/assistant";
+  isProviderWhitelisted,
+} from "@app/lib/api/assistant/models";
+import { getLargeWhitelistedModel } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
 import type { GlobalAgentSettingsModel } from "@app/lib/models/agent/agent";
 import {
