+++ front/lib/assistant.ts
@@ -6,32 +6,16 @@ import {
   isEntreprisePlanPrefix,
   isUpgraded,
 } from "@app/lib/plans/plan_codes";
-import {
-  CLAUDE_4_5_HAIKU_DEFAULT_MODEL_CONFIG,
-  CLAUDE_SONNET_4_6_DEFAULT_MODEL_CONFIG,
-} from "@app/types/assistant/models/anthropic";
-import {
-  GEMINI_2_5_FLASH_MODEL_CONFIG,
-  GEMINI_3_FLASH_MODEL_CONFIG,
-  GEMINI_3_PRO_MODEL_CONFIG,
-} from "@app/types/assistant/models/google_ai_studio";
-import {
-  MISTRAL_MEDIUM_3_5_MODEL_CONFIG,
-  MISTRAL_SMALL_MODEL_CONFIG,
-} from "@app/types/assistant/models/mistral";
-import {
-  GPT_5_5_MODEL_CONFIG,
-  GPT_5_MINI_MODEL_CONFIG,
-} from "@app/types/assistant/models/openai";
+import { CLAUDE_SONNET_4_6_DEFAULT_MODEL_CONFIG } from "@app/types/assistant/models/anthropic";
+import { GEMINI_3_PRO_MODEL_CONFIG } from "@app/types/assistant/models/google_ai_studio";
+import { MISTRAL_MEDIUM_3_5_MODEL_CONFIG } from "@app/types/assistant/models/mistral";
+import { GPT_5_5_MODEL_CONFIG } from "@app/types/assistant/models/openai";
 import { isByokProviderId } from "@app/types/assistant/models/providers";
 import type {
   ModelConfigurationType,
   ModelProviderIdType,
 } from "@app/types/assistant/models/types";
-import {
-  GROK_4_1_FAST_NON_REASONING_MODEL_CONFIG,
-  GROK_4_MODEL_CONFIG,
-} from "@app/types/assistant/models/xai";
+import { GROK_4_MODEL_CONFIG } from "@app/types/assistant/models/xai";
 import type { PlanType } from "@app/types/plan";
 import type { WhitelistableFeature } from "@app/types/shared/feature_flags";
 import type { WorkspaceType } from "@app/types/user";
@@ -40,28 +24,6 @@ export function isEnterpriseOrDust(plan: PlanType | null): boolean {
   return (
     plan !== null &&
     (isEntreprisePlanPrefix(plan.code) || isDustCompanyPlan(plan.code))
-  );
-}
-
-export function getFastestWhitelistedModel(
-  auth: Authenticator
-): ModelConfigurationType | null {
-  const whitelistedProviders = getWhitelistedProviders(auth);
-  if (whitelistedProviders.has("mistral")) {
-    return MISTRAL_SMALL_MODEL_CONFIG;
-  }
-  if (whitelistedProviders.has("google_ai_studio")) {
-    return GEMINI_2_5_FLASH_MODEL_CONFIG;
-  }
-  return _getSmallWhitelistedModel(whitelistedProviders);
-}
-
-export function getSmallWhitelistedModel(
-  auth: Authenticator,
-  excludeProviders: ReadonlySet<ModelProviderIdType> = new Set()
-): ModelConfigurationType | null {
-  return _getSmallWhitelistedModel(
-    getWhitelistedProviders(auth).difference(excludeProviders)
   );
 }
 
@@ -74,27 +36,6 @@ export function getLargeWhitelistedModel(
     getWhitelistedProviders(auth).difference(excludeProviders),
     { forBatch }
   );
-}
-
-function _getSmallWhitelistedModel(
-  whitelistedProviders: Set<ModelProviderIdType>
-): ModelConfigurationType | null {
-  if (whitelistedProviders.has("openai")) {
-    return GPT_5_MINI_MODEL_CONFIG;
-  }
-  if (whitelistedProviders.has("anthropic")) {
-    return CLAUDE_4_5_HAIKU_DEFAULT_MODEL_CONFIG;
-  }
-  if (whitelistedProviders.has("google_ai_studio")) {
-    return GEMINI_3_FLASH_MODEL_CONFIG;
-  }
-  if (whitelistedProviders.has("mistral")) {
-    return MISTRAL_SMALL_MODEL_CONFIG;
-  }
-  if (whitelistedProviders.has("xai")) {
-    return GROK_4_1_FAST_NON_REASONING_MODEL_CONFIG;
-  }
-  return null;
 }
 
 const LARGE_MODEL_CONFIGS: ModelConfigurationType[] = [
